Statistics: * Move StatisticsOneInTest to Probability package * Rename StatisticsOneInTest to ProbabilityMeasurementTest * Update ProbabilityTest - improve test case organization
